Correct Treating_Facility_ID truncation
When querying by a specific C_BioSense_Facility_ID, R is doing an under-the-hood data type conversion. It is converting Treating_Facility_ID from a string to a numeric data type. The way to get around this is by including as.is=T in your code. The as.is function changes the class of an object to indicate that it should be treated ‘as is’. This will ensure the Treating_Facility_ID is returned in its original character data type.